backend/requirements.txt
fastapi
uvicorn
jinja2
beanie
motor
python-multipart
strawberry-graphql[fastapi]
